python-fd: merge local-fileset into one file
