+++ core/platforms/freebsd/bareos-freebsd/bareos.com-common/pkg-plist.filedaemon-python-plugins-common
@@ -1,6 +1,5 @@
 
 lib/bareos/plugins/bareos-fd-local-fileset.py
 lib/bareos/plugins/BareosFdPluginBaseclass.py
-lib/bareos/plugins/BareosFdPluginLocalFileset.py
 lib/bareos/plugins/BareosFdPluginLocalFilesBaseclass.py
 lib/bareos/plugins/BareosFdWrapper.py
